chore: refactor structure
