+++ features/login/ui/index.ts
@@ -1,3 +1,3 @@
-export * from "./Circle";
-export * from "./LoginBackground";
-export * from "./LoginButton";
+export * from "@/features/login/ui/Circle";
+export * from "@/features/login/ui/LoginBackground";
+export * from "@/features/login/ui/LoginButton";
