+++ shared/ui/index.ts
@@ -1,8 +1,8 @@
-export * from "./Avatar";
-export * from "./Button";
-export * from "./CharacterBubble";
-export * from "./Chip";
-export * from "./Divider";
-export * from "./IconButton";
-export * from "./ProfileButton";
-export * from "./TextInput";
+export * from "@/shared/ui/Avatar";
+export * from "@/shared/ui/Button";
+export * from "@/shared/ui/CharacterBubble";
+export * from "@/shared/ui/Chip";
+export * from "@/shared/ui/Divider";
+export * from "@/shared/ui/IconButton";
+export * from "@/shared/ui/ProfileButton";
+export * from "@/shared/ui/TextInput";
