--- widgets/BottomNav/index.ts
@@ -1,2 +1,0 @@
-export * from "./ui";
-export * from "./model";
