+++ widgets/BottomNav/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./BottomNav";
+export * from "@/widgets/BottomNav/ui/BottomNav";
